Create SpecimenRequestManager and migrate most request methods to it from SpecimenManager
Create SimpleStudyExportContext, migrate AbstractSpecimenWriter, work to disentangle specimen writers
