+++ api/src/org/labkey/api/study/Study.java
@@ -36,6 +36,8 @@ import java.util.Map;
  */
 public interface Study extends StudyEntity
 {
+    String getShortName();
+
     List<? extends Visit> getVisits(Visit.Order order);
 
     @SuppressWarnings("unused")  // Used by cdisc_ODM StudyArchiveWriter.java
