+++ study/api-src/org/labkey/api/specimen/writer/StandardSpecimenWriter.java
@@ -16,6 +16,7 @@
 package org.labkey.api.specimen.writer;
 
 import org.labkey.api.admin.ImportContext;
+import org.labkey.api.annotations.Migrate;
 import org.labkey.api.data.ColumnInfo;
 import org.labkey.api.data.DataColumn;
 import org.labkey.api.data.DisplayColumn;
@@ -41,7 +42,8 @@ import java.util.List;
  * Date: May 7, 2009
  * Time: 9:51:36 AM
  */
-// TODO: Switch back to package-local after specimen migration? StandardSpecimenWriter needs to move to study-api.
+@Migrate
+// TODO: Switch back to package-local after specimen migration? SpecimenArchiveWriter needs to move to study-api first.
 public class StandardSpecimenWriter implements Writer<StandardSpecimenWriter.QueryInfo, ImportContext<StudyDocument.Study>>
 {
     @Override
