+++ study/src/org/labkey/study/SpecimenManager.java
@@ -24,7 +24,6 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.labkey.api.action.SpringActionController;
 import org.labkey.api.annotations.Migrate;
-import org.labkey.api.attachments.AttachmentFile;
 import org.labkey.api.attachments.AttachmentService;
 import org.labkey.api.attachments.AttachmentService.DuplicateFilenameException;
 import org.labkey.api.audit.AuditLogService;
@@ -47,10 +46,6 @@ import org.labkey.api.data.TableInfo;
 import org.labkey.api.data.TableResultSet;
 import org.labkey.api.data.TableSelector;
 import org.labkey.api.data.dialect.SqlDialect;
-import org.labkey.api.exp.Lsid;
-import org.labkey.api.exp.ObjectProperty;
-import org.labkey.api.exp.OntologyManager;
-import org.labkey.api.exp.PropertyType;
 import org.labkey.api.exp.api.ExpSampleType;
 import org.labkey.api.exp.api.ExperimentService;
 import org.labkey.api.exp.api.SampleTypeService;
@@ -59,11 +54,12 @@ import org.labkey.api.module.ModuleHtmlView;
 import org.labkey.api.query.CustomView;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.QueryService;
-import org.labkey.api.query.ValidationException;
 import org.labkey.api.security.User;
 import org.labkey.api.settings.AppProps;
+import org.labkey.api.specimen.RequestEventType;
 import org.labkey.api.specimen.SpecimenCommentAuditEvent;
 import org.labkey.api.specimen.SpecimenDetailQueryHelper;
+import org.labkey.api.specimen.SpecimenRequestManager;
 import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.specimen.SpecimenSchema;
 import org.labkey.api.specimen.SpecimenTypeBeanProperty;
@@ -71,6 +67,7 @@ import org.labkey.api.specimen.SpecimenTypeLevel;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.importer.RequestabilityManager;
 import org.labkey.api.specimen.importer.RequestabilityManager.InvalidRuleException;
+import org.labkey.api.specimen.importer.RollupHelper;
 import org.labkey.api.specimen.importer.RollupHelper.RollupMap;
 import org.labkey.api.specimen.importer.RollupInstance;
 import org.labkey.api.specimen.importer.VialSpecimenRollup;
@@ -102,7 +99,6 @@ import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.TimepointType;
 import org.labkey.api.util.DateUtil;
-import org.labkey.api.util.GUID;
 import org.labkey.api.util.HtmlString;
 import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.util.Pair;
@@ -110,7 +106,6 @@ import org.labkey.api.util.Path;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.NotFoundException;
 import org.labkey.api.view.ViewContext;
-import org.labkey.study.importer.SpecimenImporter;
 import org.labkey.study.model.CohortImpl;
 import org.labkey.study.model.StudyImpl;
 import org.labkey.study.model.StudyManager;
@@ -118,7 +113,6 @@ import org.labkey.study.model.VisitImpl;
 import org.labkey.study.query.StudyQuerySchema;
 
 import java.beans.PropertyChangeEvent;
-import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -253,60 +247,6 @@ public class SpecimenManager implements ContainerManager.ContainerListener
         return new SqlSelector(tableInfo.getSchema(), sql).getArrayList(Long.class).get(0);
     }
 
-    public List<SpecimenRequest> getRequests(Container c, User user)
-    {
-        SimpleFilter filter = new SimpleFilter(FieldKey.fromParts("Hidden"), Boolean.FALSE);
-        if (user != null)
-            filter.addCondition(FieldKey.fromParts("CreatedBy"), user.getUserId());
-        return _requestHelper.get(c, filter, "-Created");
-    }
-
-    public SpecimenRequest getRequest(Container c, int rowId)
-    {
-        return _requestHelper.get(c, rowId);
-    }
-
-    public SpecimenRequest createRequest(User user, SpecimenRequest request, boolean createEvent) throws DuplicateFilenameException
-    {
-        request = _requestHelper.create(user, request);
-        if (createEvent)
-            createRequestEvent(user, request, RequestEventType.REQUEST_CREATED, request.getRequestDescription(), null);
-        return request;
-    }
-
-    public void updateRequest(User user, SpecimenRequest request) throws InvalidRuleException
-    {
-        DbScope scope = SpecimenSchema.get().getScope();
-        try (DbScope.Transaction transaction = scope.ensureTransaction())
-        {
-            _requestHelper.update(user, request);
-
-            // update specimen states
-            List<Vial> vials = request.getVials();
-            if (vials.size() > 0)
-            {
-                SpecimenRequestStatus status = getRequestStatus(request.getContainer(), request.getStatusId());
-                updateSpecimenStatus(vials, user, status.isSpecimensLocked());
-            }
-            transaction.commit();
-        }
-    }
-
-    /**
-     * Update the lockedInRequest and available field states for the set of specimens.
-     */
-    private void updateSpecimenStatus(List<Vial> vials, User user, boolean lockedInRequest) throws InvalidRuleException
-    {
-        for (Vial vial : vials)
-        {
-            vial.setLockedInRequest(lockedInRequest);
-            Table.update(user, SpecimenSchema.get().getTableInfoVial(vial.getContainer()), vial.getRowMap(), vial.getRowId());
-        }
-        updateRequestabilityAndCounts(vials, user);
-        if (vials.size() > 0)
-            clearCaches(getContainer(vials));
-    }
-
     private Container getContainer(List<Vial> vials)
     {
         Container container = vials.get(0).getContainer();
@@ -356,7 +296,7 @@ public class SpecimenManager implements ContainerManager.ContainerListener
 
         String tableInfoSpecimenSelectName = tableInfoSpecimen.getSelectName();
         String tableInfoVialSelectName = tableInfoVial.getSelectName();
-        RollupMap<VialSpecimenRollup> matchedRollups = SpecimenImporter.getVialToSpecimenRollups(container, user);
+        RollupMap<VialSpecimenRollup> matchedRollups = RollupHelper.getVialToSpecimenRollups(container, user);
 
         SQLFragment updateSql = new SQLFragment();
         updateSql.append("UPDATE ").append(tableInfoSpecimenSelectName).append(UPDATE_SPECIMEN_SETS);
@@ -453,31 +393,6 @@ public class SpecimenManager implements ContainerManager.ContainerListener
         if (request == null)
             return new SpecimenRequestRequirement[0];
         return request.getRequirements();
-    }
-
-    public void deleteRequestRequirement(User user, SpecimenRequestRequirement requirement) throws DuplicateFilenameException
-    {
-        deleteRequestRequirement(user, requirement, true);
-    }
-
-    public void deleteRequestRequirement(User user, SpecimenRequestRequirement requirement, boolean createEvent) throws DuplicateFilenameException
-    {
-        if (createEvent)
-            createRequestEvent(user, requirement, RequestEventType.REQUIREMENT_REMOVED, requirement.getRequirementSummary(), null);
-        requirement.delete();
-    }
-
-    public void createRequestRequirement(User user, SpecimenRequestRequirement requirement, boolean createEvent) throws DuplicateFilenameException
-    {
-        createRequestRequirement(user, requirement, createEvent, false);
-    }
-
-    public void createRequestRequirement(User user, SpecimenRequestRequirement requirement, boolean createEvent, boolean force) throws DuplicateFilenameException
-    {
-        SpecimenRequest request = getRequest(requirement.getContainer(), requirement.getRequestId());
-        SpecimenRequestRequirement newRequirement = SpecimenRequestRequirementProvider.get().createRequirement(user, request, requirement, force);
-        if (newRequirement != null && createEvent)
-            createRequestEvent(user, requirement, RequestEventType.REQUIREMENT_ADDED, requirement.getRequirementSummary(), null);
     }
 
     public void updateRequestRequirement(User user, SpecimenRequestRequirement requirement)
@@ -633,21 +548,6 @@ public class SpecimenManager implements ContainerManager.ContainerListener
         return uniqueStatuses;
     }
 
-    public void createRequestStatus(User user, SpecimenRequestStatus status)
-    {
-        _requestStatusHelper.create(user, status);
-    }
-
-    public void updateRequestStatus(User user, SpecimenRequestStatus status)
-    {
-        _requestStatusHelper.update(user, status);
-    }
-
-    public void deleteRequestStatus(User user, SpecimenRequestStatus status)
-    {
-        _requestStatusHelper.delete(status);
-    }
-
     public List<SpecimenRequestEvent> getRequestEvents(Container c)
     {
         return _requestEventHelper.get(c);
@@ -656,73 +556,6 @@ public class SpecimenManager implements ContainerManager.ContainerListener
     public SpecimenRequestEvent getRequestEvent(Container c, int rowId)
     {
         return _requestEventHelper.get(c, rowId);
-    }
-
-    public enum RequestEventType
-    {
-        REQUEST_CREATED("Request Created"),
-        REQUEST_STATUS_CHANGED("Request Status Changed"),
-        REQUIREMENT_ADDED("Requirement Created"),
-        REQUIREMENT_REMOVED("Requirement Removed"),
-        REQUIREMENT_UPDATED("Requirement Updated"),
-        REQUEST_UPDATED("Request Updated"),
-        SPECIMEN_ADDED("Specimen Added"),
-        SPECIMEN_REMOVED("Specimen Removed"),
-        SPECIMEN_LIST_GENERATED("Specimen List Generated"),
-        COMMENT_ADDED("Comment/Attachment(s) Added"),
-        NOTIFICATION_SENT("Notification Sent");
-
-        private final String _displayText;
-
-        RequestEventType(String displayText)
-        {
-            _displayText = displayText;
-        }
-
-        public String getDisplayText()
-        {
-            return _displayText;
-        }
-    }
-
-    public SpecimenRequestEvent createRequestEvent(User user, SpecimenRequestRequirement requirement, RequestEventType type, String comments, List<AttachmentFile> attachments) throws DuplicateFilenameException
-    {
-        return createRequestEvent(user, requirement.getContainer(), requirement.getRequestId(), requirement.getRowId(), type, comments, attachments);
-    }
-
-    public SpecimenRequestEvent createRequestEvent(User user, SpecimenRequest request, RequestEventType type, String comments, List<AttachmentFile> attachments) throws DuplicateFilenameException
-    {
-        return createRequestEvent(user, request.getContainer(), request.getRowId(), -1, type, comments, attachments);
-    }
-
-    private SpecimenRequestEvent createRequestEvent(User user, Container container, int requestId, int requirementId, RequestEventType type, String comments, List<AttachmentFile> attachments) throws DuplicateFilenameException
-    {
-        SpecimenRequestEvent event = new SpecimenRequestEvent();
-        event.setEntryType(type.getDisplayText());
-        event.setComments(comments);
-        event.setRequestId(requestId);
-        event.setCreated(new Date(System.currentTimeMillis()));
-        if (requirementId >= 0)
-            event.setRequirementId(requirementId);
-        event.setContainer(container);
-        event.setEntityId(GUID.makeGUID());
-        event = createRequestEvent(user, event);
-        try
-        {
-            AttachmentService.get().addAttachments(event, attachments, user);
-        }
-        catch (DuplicateFilenameException e)
-        {
-            // UI should (minimally) catch and display these errors nicely
-            throw e;
-        }
-        catch (IOException e)
-        {
-            // this is unexpected, and indicative of a larger system problem; we'll convert to a runtime
-            // exception, rather than requiring all event loggers to handle this unlikely scenario:
-            throw new RuntimeException(e);
-        }
-        return event;
     }
 
     private void deleteRequestEvents(User user, SpecimenRequest request)
@@ -734,11 +567,6 @@ public class SpecimenManager implements ContainerManager.ContainerListener
             AttachmentService.get().deleteAttachments(event);
             _requestEventHelper.delete(event);
         }
-    }
-
-    private SpecimenRequestEvent createRequestEvent(User user, SpecimenRequestEvent event)
-    {
-        return _requestEventHelper.create(user, event);
     }
 
     @NotNull
@@ -768,259 +596,6 @@ public class SpecimenManager implements ContainerManager.ContainerListener
         return vials;
     }
 
-    public static class SpecimenRequestInput
-    {
-        private final String _title;
-        private final String _helpText;
-        private final int _displayOrder;
-
-        private boolean _required;
-        private boolean _rememberSiteValue;
-        private boolean _multiLine;
-        private Map<Integer,String> _locationToDefaultValue;
-
-        public SpecimenRequestInput(String title, String helpText, int displayOrder, boolean multiLine, boolean required, boolean rememberSiteValue)
-        {
-            _title = title;
-            _required = required;
-            _rememberSiteValue = rememberSiteValue;
-            _helpText = helpText;
-            _displayOrder = displayOrder;
-            _multiLine = multiLine;
-        }
-
-        public SpecimenRequestInput(String title, String helpText, int displayOrder)
-        {
-            this(title, helpText, displayOrder, false, false, false);
-        }
-
-        public String getHelpText()
-        {
-            return _helpText;
-        }
-
-        public boolean isRememberSiteValue()
-        {
-            return _rememberSiteValue;
-        }
-
-        public boolean isRequired()
-        {
-            return _required;
-        }
-
-        public String getTitle()
-        {
-            return _title;
-        }
-
-        public int getDisplayOrder()
-        {
-            return _displayOrder;
-        }
-
-        public boolean isMultiLine()
-        {
-            return _multiLine;
-        }
-
-        public void setMultiLine(boolean multiLine)
-        {
-            _multiLine = multiLine;
-        }
-
-        public void setRememberSiteValue(boolean rememberSiteValue)
-        {
-            _rememberSiteValue = rememberSiteValue;
-        }
-
-        public void setRequired(boolean required)
-        {
-            _required = required;
-        }
-
-        public Map<Integer,String> getDefaultSiteValues(Container container) throws ValidationException
-        {
-            if (!isRememberSiteValue())
-                throw new UnsupportedOperationException("Only those inputs set to remember site values can be queried for a site default.");
-
-            if (_locationToDefaultValue != null)
-                return _locationToDefaultValue;
-            String defaultObjectLsid = getRequestInputDefaultObjectLsid(container);
-            String setItemLsid = ensureOntologyManagerSetItem(container, defaultObjectLsid, getTitle());
-            Map<Integer, String> locationToValue = new HashMap<>();
-
-            Map<String, ObjectProperty> defaultValueProperties = OntologyManager.getPropertyObjects(container, setItemLsid);
-            if (defaultValueProperties != null)
-            {
-                for (Map.Entry<String, ObjectProperty> defaultValue : defaultValueProperties.entrySet())
-                {
-                    String locationIdString = defaultValue.getKey().substring(defaultValue.getKey().lastIndexOf(".") + 1);
-                    int locationId = Integer.parseInt(locationIdString);
-                    locationToValue.put(locationId, defaultValue.getValue().getStringValue());
-                }
-            }
-            _locationToDefaultValue = locationToValue;
-            return _locationToDefaultValue;
-        }
-
-        public void setDefaultSiteValue(Container container, int locationId, String value) throws SQLException
-        {
-            try
-            {
-                assert locationId > 0 : "Invalid site id: " + locationId;
-                if (!isRememberSiteValue())
-                    throw new UnsupportedOperationException("Only those inputs configured to remember site values can set a site default.");
-                _locationToDefaultValue = null;
-                String parentObjectLsid = getRequestInputDefaultObjectLsid(container);
-
-                String setItemLsid = ensureOntologyManagerSetItem(container, parentObjectLsid, getTitle());
-                String propertyId = parentObjectLsid + "." + locationId;
-                ObjectProperty defaultValueProperty = new ObjectProperty(setItemLsid, container, propertyId, value);
-                OntologyManager.deleteProperty(setItemLsid, propertyId, container, container);
-                OntologyManager.insertProperties(container, setItemLsid, defaultValueProperty);
-            }
-            catch (ValidationException e)
-            {
-                throw new SQLException(e.getMessage());
-            }
-        }
-    }
-
-    public SpecimenRequestInput[] getNewSpecimenRequestInputs(Container container) throws SQLException
-    {
-        return getNewSpecimenRequestInputs(container, true);
-    }
-
-    public SpecimenRequestInput[] getNewSpecimenRequestInputs(Container container, boolean createIfMissing) throws SQLException
-    {
-        String parentObjectLsid = getRequestInputObjectLsid(container);
-        Map<String,ObjectProperty> resourceProperties = OntologyManager.getPropertyObjects(container, parentObjectLsid);
-        SpecimenRequestInput[] inputs = new SpecimenRequestInput[0];
-        if (resourceProperties == null || resourceProperties.size() == 0)
-        {
-            if (createIfMissing)
-            {
-                try (var ignore = SpringActionController.ignoreSqlUpdates())
-                {
-                    inputs = new SpecimenRequestInput[] {
-                            new SpecimenRequestInput("Assay Plan", "Please enter a description of or reference to the assay plan(s) that will be used for the requested specimens.", 0, true, true, false),
-                            new SpecimenRequestInput("Shipping Information", "Please enter your shipping address along with any special instructions.", 1, true, true, true),
-                            new SpecimenRequestInput("Comments", "Please enter any additional information regarding your request.", 2, true, false, false)
-                    };
-                    saveNewSpecimenRequestInputs(container, inputs);
-                }
-            }
-            return inputs;
-        }
-        else
-        {
-            inputs = new SpecimenRequestInput[resourceProperties.size()];
-            for (Map.Entry<String, ObjectProperty> parentPropertyEntry : resourceProperties.entrySet())
-            {
-                String resourcePropertyLsid = parentPropertyEntry.getKey();
-                int displayOrder = Integer.parseInt(resourcePropertyLsid.substring(resourcePropertyLsid.lastIndexOf('.') + 1));
-
-                Map<String, ObjectProperty> inputProperties = parentPropertyEntry.getValue().retrieveChildProperties();
-                String title = inputProperties.get(parentObjectLsid + ".Title").getStringValue();
-                String helpText = null;
-                if (inputProperties.get(parentObjectLsid + ".HelpText") != null)
-                    helpText = inputProperties.get(parentObjectLsid + ".HelpText").getStringValue();
-                boolean rememberSiteValue = inputProperties.get(parentObjectLsid + ".RememberSiteValue").getFloatValue() == 1;
-                boolean required = inputProperties.get(parentObjectLsid + ".Required").getFloatValue() == 1;
-                boolean multiLine = inputProperties.get(parentObjectLsid + ".MultiLine").getFloatValue() == 1;
-                inputs[displayOrder] = new SpecimenRequestInput(title, helpText, displayOrder, multiLine, required, rememberSiteValue);
-            }
-        }
-        return inputs;
-    }
-
-    private static String getRequestInputObjectLsid(Container container)
-    {
-        return new Lsid(StudyService.SPECIMEN_NAMESPACE_PREFIX, "Folder-" + container.getRowId(), "RequestInput").toString();
-    }
-
-    private static String getRequestInputDefaultObjectLsid(Container container)
-    {
-        return new Lsid(StudyService.SPECIMEN_NAMESPACE_PREFIX, "Folder-" + container.getRowId(), "RequestInputDefault").toString();
-    }
-
-    private static String ensureOntologyManagerSetItem(Container container, String lsidBase, String uniqueItemId) throws ValidationException
-    {
-        try (var ignore = SpringActionController.ignoreSqlUpdates())
-        {
-            Integer listParentObjectId = OntologyManager.ensureObject(container, lsidBase);
-            String listItemReferenceLsidPrefix = lsidBase + "#objectResource.";
-            String listItemObjectLsid = lsidBase + "#" + uniqueItemId;
-            String listItemPropertyReferenceLsid = listItemReferenceLsidPrefix + uniqueItemId;
-
-            // ensure the object that corresponds to a single list item:
-            OntologyManager.ensureObject(container, listItemObjectLsid, listParentObjectId);
-
-            // check to make sure that the list item is wired up to the top-level list object via a property:
-            Map<String, ObjectProperty> properties = OntologyManager.getPropertyObjects(container, lsidBase);
-            if (!properties.containsKey(listItemPropertyReferenceLsid))
-            {
-                // create the resource property that links the parent object to the list item object:
-                ObjectProperty resourceProperty = new ObjectProperty(lsidBase, container,
-                        listItemPropertyReferenceLsid, listItemObjectLsid, PropertyType.RESOURCE);
-                OntologyManager.insertProperties(container, lsidBase, resourceProperty);
-            }
-            return listItemObjectLsid;
-        }
-    }
-
-    public void saveNewSpecimenRequestInputs(Container container, SpecimenRequestInput[] inputs) throws SQLException
-    {
-        if (!requestInputsChanged(container, inputs))
-            return;
-
-        try {
-            String parentObjectLsid = getRequestInputObjectLsid(container);
-            String defaultValuesObjectLsid = getRequestInputDefaultObjectLsid(container);
-            OntologyManager.deleteOntologyObject(parentObjectLsid, container, true);
-            OntologyManager.deleteOntologyObject(defaultValuesObjectLsid, container, true);
-            for (int i = 0; i < inputs.length; i++)
-            {
-                SpecimenRequestInput input = inputs[i];
-                String setItemLsid = ensureOntologyManagerSetItem(container, parentObjectLsid, "" + i);
-                ObjectProperty[] props = new ObjectProperty[5];
-                props[0] = new ObjectProperty(setItemLsid, container, parentObjectLsid + ".HelpText", input.getHelpText());
-                props[1] = new ObjectProperty(setItemLsid, container, parentObjectLsid + ".Required", input.isRequired() ? 1 : 0);
-                props[2] = new ObjectProperty(setItemLsid, container, parentObjectLsid + ".RememberSiteValue", input.isRememberSiteValue() ? 1 : 0);
-                props[3] = new ObjectProperty(setItemLsid, container, parentObjectLsid + ".Title", input.getTitle());
-                props[4] = new ObjectProperty(setItemLsid, container, parentObjectLsid + ".MultiLine", input.isMultiLine() ? 1 : 0);
-                OntologyManager.insertProperties(container, setItemLsid, props);
-            }
-        }
-        catch (ValidationException ve)
-        {
-            throw new SQLException(ve.getMessage());
-        }
-    }
-
-    private boolean requestInputsChanged(Container container, SpecimenRequestInput[] newInputs) throws SQLException
-    {
-        SpecimenRequestInput[] oldInputs = getNewSpecimenRequestInputs(container, false);
-        if (oldInputs.length == 0)
-            return true;
-        else if (oldInputs.length != newInputs.length)
-            return true;
-        else
-        {
-            for (int i = 0; i < oldInputs.length; i++)
-            {
-                if (oldInputs[i].isMultiLine() != newInputs[i].isMultiLine() ||
-                    oldInputs[i].isRememberSiteValue() != newInputs[i].isRememberSiteValue() ||
-                    oldInputs[i].isRequired() != newInputs[i].isRequired() ||
-                    !oldInputs[i].getTitle().equals(newInputs[i].getTitle()) ||
-                    !getSafeString(oldInputs[i].getHelpText()).equals(getSafeString(newInputs[i].getHelpText())))
-                    return true;
-            }
-        }
-        return false;
-    }
-
     private String getSafeString(String str)
     {
         if (str == null)
@@ -1030,6 +605,7 @@ public class SpecimenManager implements ContainerManager.ContainerListener
     }
 
     private static final ReentrantLock REQUEST_ADDITION_LOCK = new ReentrantLock();
+
     public void createRequestSpecimenMapping(User user, SpecimenRequest request, List<Vial> vials, boolean createEvents, boolean createRequirements)
             throws InvalidRuleException, DuplicateFilenameException, SpecimenRequestException
     {
@@ -1056,7 +632,7 @@ public class SpecimenManager implements ContainerManager.ContainerListener
                 fields.put("SpecimenGlobalUniqueId", vial.getGlobalUniqueId());
                 Table.insert(user, SpecimenSchema.get().getTableInfoSampleRequestSpecimen(), fields);
                 if (createEvents)
-                    createRequestEvent(user, request, RequestEventType.SPECIMEN_ADDED, vial.getSpecimenDescription(), null);
+                    SpecimenRequestManager.get().createRequestEvent(user, request, RequestEventType.SPECIMEN_ADDED, vial.getSpecimenDescription(), null);
             }
 
             if (createRequirements)
@@ -1067,6 +643,21 @@ public class SpecimenManager implements ContainerManager.ContainerListener
 
             transaction.commit();
         }
+    }
+
+    /**
+     * Update the lockedInRequest and available field states for the set of specimens.
+     */
+    private void updateSpecimenStatus(List<Vial> vials, User user, boolean lockedInRequest) throws RequestabilityManager.InvalidRuleException
+    {
+        for (Vial vial : vials)
+        {
+            vial.setLockedInRequest(lockedInRequest);
+            Table.update(user, SpecimenSchema.get().getTableInfoVial(vial.getContainer()), vial.getRowMap(), vial.getRowId());
+        }
+        updateRequestabilityAndCounts(vials, user);
+        if (vials.size() > 0)
+            clearCaches(getContainer(vials));
     }
 
     public List<Vial> getVials(Container container, User user, int[] vialsRowIds)
@@ -1177,7 +768,7 @@ public class SpecimenManager implements ContainerManager.ContainerListener
             if (createEvents)
             {
                 for (String description : descriptions)
-                    createRequestEvent(user, request, RequestEventType.SPECIMEN_REMOVED, description, null);
+                    SpecimenRequestManager.get().createRequestEvent(user, request, RequestEventType.SPECIMEN_REMOVED, description, null);
             }
 
             updateSpecimenStatus(vials, user, false);
