+++ study/src/org/labkey/study/controllers/CreateChildStudyAction.java
@@ -34,6 +34,7 @@ import org.labkey.api.query.ValidationException;
 import org.labkey.api.security.RequiresPermission;
 import org.labkey.api.security.User;
 import org.labkey.api.security.permissions.AdminPermission;
+import org.labkey.api.specimen.SpecimenRequestManager;
 import org.labkey.api.specimen.SpecimenSchema;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.requirements.SpecimenRequest;
@@ -168,7 +169,7 @@ public class CreateChildStudyAction extends MutatingApiAction<ChildStudyDefiniti
 
             if (null != form.getRequestId())
             {
-                SpecimenRequest request = sm.getRequest(sourceContainer, form.getRequestId());
+                SpecimenRequest request = SpecimenRequestManager.get().getRequest(sourceContainer, form.getRequestId());
 
                 if (null == request)
                 {
