+++ study/src/org/labkey/study/importer/EditableSpecimenImporter.java
@@ -29,6 +29,7 @@ import org.labkey.api.security.User;
 import org.labkey.api.specimen.SpecimenSchema;
 import org.labkey.api.specimen.importer.EventVialRollup;
 import org.labkey.api.specimen.importer.IteratorSpecimenImportFile;
+import org.labkey.api.specimen.importer.RollupHelper;
 import org.labkey.api.specimen.importer.RollupHelper.RollupMap;
 import org.labkey.api.specimen.importer.RollupInstance;
 import org.labkey.api.specimen.importer.SpecimenColumn;
@@ -194,7 +195,7 @@ public class EditableSpecimenImporter extends SpecimenImporter
             _specialColumnNameMap.put("latestqualitycomments", "qualitycomments");
 
             // Add any rollups from optionals whose names are not matching
-            RollupMap<EventVialRollup> matchedRollups = SpecimenImporter.getEventToVialRollups(getContainer(), getUser());
+            RollupMap<EventVialRollup> matchedRollups = RollupHelper.getEventToVialRollups(getContainer(), getUser());
             for (Map.Entry<String, List<RollupInstance<EventVialRollup>>> entry : matchedRollups.entrySet())
             {
                 String fromName = entry.getKey();
