+++ study/src/org/labkey/study/view/specimen/manageRequestInputs.jsp
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 %>
+<%@ page import="org.labkey.api.specimen.SpecimenRequestManager.SpecimenRequestInput"%>
 <%@ page import="org.labkey.api.study.StudyUrls"%>
 <%@ page import="org.labkey.api.view.HttpView"%>
 <%@ page import="org.labkey.api.view.JspView"%>
-<%@ page import="org.labkey.study.SpecimenManager"%>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.HandleUpdateRequestInputsAction"%>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.ManageRequestInputsBean" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
@@ -179,10 +179,10 @@
                     "If checked, the input will be pre-populated with the previous value entered for the destination location.")%></th>
         </tr>
     <%
-        SpecimenManager.SpecimenRequestInput[] inputs = bean.getInputs();
+        SpecimenRequestInput[] inputs = bean.getInputs();
         for (int inputIndex = 0; inputIndex < inputs.length; inputIndex++)
         {
-            SpecimenManager.SpecimenRequestInput input = inputs[inputIndex];
+            SpecimenRequestInput input = inputs[inputIndex];
     %>
         <tr>
             <td style="padding-right: 5px;"><%= text(tdButtons) %></td>
