+++ study/src/org/labkey/study/view/specimen/manageStatuses.jsp
@@ -15,11 +15,12 @@
  * limitations under the License.
  */
 %>
+<%@ page import="org.labkey.api.specimen.SpecimenRequestManager"%>
 <%@ page import="org.labkey.api.specimen.SpecimenRequestStatus"%>
 <%@ page import="org.labkey.api.specimen.settings.SettingsManager"%>
 <%@ page import="org.labkey.api.specimen.settings.StatusSettings"%>
 <%@ page import="org.labkey.api.study.StudyUrls"%>
-<%@ page import="org.labkey.api.view.ActionURL"%>
+<%@ page import="org.labkey.api.view.ActionURL" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.DeleteStatusAction" %>
@@ -33,7 +34,7 @@
 <%
     JspView<StudyImpl> me = (JspView<StudyImpl>) HttpView.currentView();
     StudyImpl study = me.getModelBean();
-    List<SpecimenRequestStatus> statuses = study.getSampleRequestStatuses(getUser());
+    List<SpecimenRequestStatus> statuses = SpecimenRequestManager.get().getRequestStatuses(study.getContainer(), getUser());
     Set<Integer> inUseStatuses = study.getSampleRequestStatusesInUse();
     StatusSettings settings = SettingsManager.get().getStatusSettings(study.getContainer());
     boolean showSystemStatuses = settings.isUseShoppingCart();
